refactor(shell): extract wingman alerts and screenshot
Split shell/js/wingman/index.js:
- alerts.js: transient overlay toasts + window.dismissAlert binding
- screenshot.js: screenshot button + region-capture overlay

window.dismissAlert remains bound for inline onclick in index.html.
No logic changes.
